requirements.txt
absl-py==0.11.0
appdirs==1.4.4
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cmake==3.18.2.post1
cycler==0.10.0
distlib==0.3.1
dlib==19.21.0
docutils==0.16
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.0.12
firebase==3.0.1
gast==0.3.3
gcloud==0.17.0
google-api-core==1.24.1
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-cloud==0.34.0
google-cloud-core==1.5.0
google-cloud-storage==1.35.0
google-crc32c==1.1.0
google-pasta==0.2.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.33.2
h5py==2.10.0
httplib2==0.18.1
idna==2.10
importlib-metadata==2.0.0
imutils==0.5.3
joblib==0.17.0
jws==0.1.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.3.3
matplotlib==3.3.2
numpy==1.18.4
oauth2client==3.0.0
oauthlib==3.1.0
opencv-contrib-python==4.4.0.44
opt-einsum==3.3.0
pandas==1.1.3
Pillow==7.2.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.4.3
pyparsing==2.4.7
Pyrebase==3.0.27
python-dateutil==2.8.1
python-jwt==2.0.1
pytz==2020.1
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.7.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.4.1
seaborn==0.11.0
six==1.15.0
sklearn==0.0
statistics==1.0.3.5
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow-addons==0.11.2
tensorflow-cpu==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
typeguard==2.10.0
urllib3==1.25.10
virtualenv==20.0.33
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.3.0
